+++ src/components/tags/Search.vue
@@ -15,7 +15,7 @@
               v-focus>
           </div>
           <div class="control">
-            <a class="button is-static">{{ totalCount }}</a>
+            <a class="button is-static">{{ totalCount | formatNumber }}</a>
           </div>
         </div>